src/main/java/guru/springframework/sfgrecipeproject/controllers/RecipeController.java
package guru.springframework.sfgrecipeproject.controllers;

import guru.springframework.sfgrecipeproject.domain.Recipe;
import guru.springframework.sfgrecipeproject.service.RecipeService;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.List;

@Controller
@RequestMapping("/recipe")
public class RecipeController {
    private final RecipeService recipeService;

    public RecipeController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @GetMapping({"/list", "list.html"})
    public String recipeList(Model model) {

        List<Recipe> recipes = recipeService.getRecipes();
        model.addAttribute("recipes", recipes);

        return "recipe/list";
    }

    @GetMapping("view/{id}")
    public String recipeView(@PathVariable String id, Model model) {

        Recipe recipe = recipeService.findById(Long.valueOf(id));
        model.addAttribute("recipe", recipe);

        return "recipe/view";
    }
}
